@@ -25,8 +25,8 @@ package org.nhindirect.stagent;
 import java.util.ArrayList;
 import java.util.Collection;
 
-import javax.mail.internet.AddressException;
-import javax.mail.internet.InternetAddress;
+import jakarta.mail.internet.AddressException;
+import jakarta.mail.internet.InternetAddress;
 
 import org.nhindirect.stagent.trust.TrustEnforcementStatus;
 
