@@ -4,7 +4,7 @@ import java.io.File;
 import java.security.cert.X509Certificate;
 import java.util.Collection;
 
-import javax.mail.internet.InternetAddress;
+import jakarta.mail.internet.InternetAddress;
 
 import org.apache.commons.io.FileUtils;
 import org.nhindirect.stagent.cert.impl.LDAPCertificateStore;
