@@ -1,7 +1,7 @@
 package org.nhindirect.stagent.mail.notifications;
 
-import javax.mail.internet.ContentType;
-import javax.mail.internet.MimeMessage;
+import jakarta.mail.internet.ContentType;
+import jakarta.mail.internet.MimeMessage;
 
 import org.nhindirect.stagent.mail.MailStandard;
 import org.nhindirect.stagent.mail.MimeEntity;
