@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Collection;
 
-import javax.mail.internet.InternetAddress;
+import jakarta.mail.internet.InternetAddress;
 
 import org.nhindirect.policy.PolicyExpression;
 
