@@ -10,7 +10,7 @@ import java.io.ByteArrayInputStream;
 import java.security.cert.X509Certificate;
 import java.util.Arrays;
 
-import javax.mail.internet.InternetAddress;
+import jakarta.mail.internet.InternetAddress;
 
 import org.bouncycastle.cms.CMSSignedData;
 import org.nhindirect.common.crypto.CryptoExtensions;
