@@ -67,8 +67,7 @@ public class CRLRevocationManager_getCRLCollectionTest
 		String uri = "http://localhost:8080/master.crl";
 		
 		
-		SoftReference<X509CRL> softRef = mock(SoftReference.class);
-		when(softRef.get()).thenReturn(null);
+		SoftReference<X509CRL> softRef = new SoftReference<>(null);
 		
 		CRLRevocationManager.cache.put(uri, softRef);
 		
