@@ -103,14 +103,12 @@ public class CRLRevocationManager_getCrlFromUriTest
 	}
 	
 	@Test
-	@SuppressWarnings("unchecked")
 	public void testGetCrlFromUri_existsInCache_softRefRemoved_assertCRLNotFound()
 	{
 		String uri = "http://localhost:8080/master.crl";
 		
 		
-		SoftReference<X509CRL> softRef = mock(SoftReference.class);
-		when(softRef.get()).thenReturn(null);
+		SoftReference<X509CRL> softRef = new SoftReference<>(null);
 		
 		CRLRevocationManager.cache.put(uri, softRef);
 		
